Atomically run `make install` against whatever version was built
If (in the extremely unlikely event!) that a new version of nginx is
released in between the time that nginx was built and `make install`
were run, version.txt would have been updated again. Force `make
install` to go with whatever version.txt says, even if it's out of date
(can also be used to manually roll back version.txt to a version that
was previously downloaded and processed).